fix: use topics and move rest implementations to rest.rs
